fix: add exists
