+++ package.json
@@ -2,10 +2,10 @@
   "name": "fm",
   "version": "1.0.0",
   "description": "",
-  "main": "src/index.js",
+  "main": "src/index.mjs",
   "type": "module",
   "scripts": {
-    "start": "node src/index.js"
+    "start": "node src/index.mjs"
   },
   "author": "",
   "license": "ISC",
